Last stage commit

+++ src/LineDataObject.java
@@ -4,5 +4,5 @@ import java.util.TreeMap;
 
 public class LineDataObject {
 	Map<String, Integer> lineMap = new TreeMap<String, Integer>();
-	int uniqueWords;
+	double uniqueWords;
 }
